--- a/core-relations/src/action/mask.rs
+++ b/core-relations/src/action/mask.rs
@@ -101,6 +101,10 @@ impl Mask {
     pub(super) fn clear(&mut self) {
         self.data.clear();
     }
+
+    pub(crate) fn count_ones(&self) -> usize {
+        self.data.count_ones(..)
+    }
 }
 
 pub(crate) enum IterResult<T> {