--- a/core-relations/src/action/mod.rs
+++ b/core-relations/src/action/mod.rs
@@ -517,7 +517,7 @@ impl<'a> ExecutionState<'a> {
 }
 
 impl ExecutionState<'_> {
-    pub(crate) fn run_instrs(&mut self, instrs: &[Instr], bindings: &mut Bindings) {
+    pub(crate) fn run_instrs(&mut self, instrs: &[Instr], bindings: &mut Bindings) -> usize {
         if bindings.var_offsets.next_id().rep() == 0 {
             // If we have no variables, we want to run the rules once.
             bindings.matches = 1;
@@ -527,10 +527,11 @@ impl ExecutionState<'_> {
         let mut mask = with_pool_set(|ps| Mask::new(0..bindings.matches, ps));
         for instr in instrs {
             if mask.is_empty() {
-                break;
+                return 0;
             }
             self.run_instr(&mut mask, instr, bindings);
         }
+        return mask.count_ones();
     }
     fn run_instr(&mut self, mask: &mut Mask, inst: &Instr, bindings: &mut Bindings) {
         fn assert_impl(