The same description may correspond to multiple plans.