@yihozhang 2025-10-05
The new variable ordering makes python-array-optimize slower:


This PR:
```
Ruleset combined_ruleset_4774957744: search 0.049s, merge 0.254s, rebuild 0.442s
Ruleset combined_ruleset_4774954096: search 0.004s, merge 0.021s, rebuild 0.056s
Ruleset combined_ruleset_4774912912: search 1.180s, merge 2.009s, rebuild 2.275s
```

Current main (`e28e4e0`):
```
Ruleset combined_ruleset_4774957744: search 0.047s, merge 0.258s, rebuild 0.446s
Ruleset combined_ruleset_4774954096: search 0.004s, merge 0.022s, rebuild 0.057s
Ruleset combined_ruleset_4774912912: search 0.558s, merge 1.959s, rebuild 2.230s
```