@ezrosent 2025-10-08
I think it's worth checking with Saul on whether we should gate old vs. new ordering via a flag or something to avoid hitting python performance more, depending on how representative python-array-optimize is. I know the new backend is already a bit slower than 0.5 so pesimizing further isn't great. But up to y'all; I'm good with merging.