@saulshanabrook 2025-10-08
Yeah I wouldn't block on the performance slowdown here. It's bad, but not like orders of magnitude worse. I think the benchmark did it's job of highlighting the issue, it's OK if we choose to move on for the sake of code clarity/ease, if there isn't an elegant way to address it right now or that would take too much effort.